# BackendEngineering/task-2/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A simple CRUD Task API built with Node.js and Express."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "description": "Returns a list of all tasks.",
        "responses": {
          "200": {
            "description": "A list of tasks."
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "description": "Creates a new task with the provided title.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "title": "Buy milk"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully."
          },
          "400": {
            "description": "Title is missing."
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get a task by ID",
        "description": "Returns a single task.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task found."
          },
          "404": {
            "description": "Task not found."
          }
        }
      },
      "put": {
        "summary": "Update a task",
        "description": "Updates a task's title and/or done status.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "title": "Learn Swagger",
                "done": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully."
          },
          "400": {
            "description": "Invalid request body."
          },
          "404": {
            "description": "Task not found."
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "description": "Deletes a task by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted successfully."
          },
          "404": {
            "description": "Task not found."
          }
        }
      }
    }
  }
}